[Refactor] Tuist Configuartion, Info 관련 구조 변경
- 각 모듈별이 아닌 app의 configuration, info 값들을 가져와서 사용하도록 수정함
